backend/requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-jsonapi==7.0.2
djangorestframework-simplejwt==5.3.1
idna==3.7
inflection==0.5.1
mysqlclient==2.2.4
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.1
urllib3==2.2.2
